Fixes errors in ArraySource tests.
Since CakePHP 2.3 the test was throwing errors due to a change in
the core. The change was related to the order clause of the query.

Previously CakePHP passed the order clause as an array of compiled
strings. Since 2.3 CakePHP passes multidimensional arrays when
they are defined as such in the original query.

This core change caused implode() to throw errors because it can't
handle multidimensional arrays.